Add files via upload
We want to maximize the sum of the cells in the top-left quadrant. For each cell in the final configuration of the top-left quadrant, there are candidates, meaning the corresponding cells in the top-left, top-right, bottom-left, and bottom-right submatrices that it can be swapped with.

We have to show that for each cell, wherever it is, we can swap it with the corresponding candidate value in the top-left submatrix without changing the order of the other cells in the top-left submatrix. The diagrams below demonstrate this for an instance where the maximum value of the candidates is in the top-right submatrix. If it is in the bottom-left or bottom-right submatrices, you can first reverse a row and/or column to put it in the top-right submatrix and then follow the same sequence of operations shown in the diagram below.